+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_completion/domain/analytic/StepCompletionDailyStepCompletedModalClickedGoBackHyperskillAnalyticEvent.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.step_completion.domain.analytic.topic_completed_modal
+package org.hyperskill.app.step_completion.domain.analytic
 
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
@@ -7,24 +7,24 @@ import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRou
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticTarget
 
 /**
- * Represents click on the "Go to homescreen" button analytic event.
+ * Represents click on the "Go back" button analytic event.
  *
  * JSON payload:
  * ```
  * {
  *     "route": "/learn/step/1",
  *     "action": "click",
- *     "part": "topic_completed_modal",
- *     "target": "go_to_home_screen"
+ *     "part": "daily_step_completed_modal",
+ *     "target": "go_back"
  * }
  * ```
  * @see HyperskillAnalyticEvent
  */
-class StepCompletionTopicCompletedModalClickedGoToHomeScreenHyperskillAnalyticEvent(
+class StepCompletionDailyStepCompletedModalClickedGoBackHyperskillAnalyticEvent(
     route: HyperskillAnalyticRoute
 ) : HyperskillAnalyticEvent(
     route,
     HyperskillAnalyticAction.CLICK,
-    HyperskillAnalyticPart.TOPIC_COMPLETED_MODAL,
-    HyperskillAnalyticTarget.GO_TO_HOME_SCREEN
+    HyperskillAnalyticPart.DAILY_STEP_COMPLETED_MODAL,
+    HyperskillAnalyticTarget.GO_BACK
 )
