+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_completion/domain/analytic/StepCompletionDailyStepCompletedModalShownHyperskillAnalyticEvent.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.step_completion.domain.analytic.daily_notifications_notice
+package org.hyperskill.app.step_completion.domain.analytic
 
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
@@ -7,24 +7,24 @@ import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRou
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticTarget
 
 /**
- * Represents a shown analytic event of the prompt to receive daily study reminders.
+ * Represents a shown analytic event of the daily step completed modal.
  *
  * JSON payload:
  * ```
  * {
  *     "route": "/learn/step/1",
- *     "action": "hidden",
- *     "part": "notice",
- *     "target": "daily_notifications_notice"
+ *     "action": "shown",
+ *     "part": "modal",
+ *     "target": "daily_step_completed_modal"
  * }
  * ```
  * @see HyperskillAnalyticEvent
  */
-class StepCompletionShownDailyNotificationsNoticeHyperskillAnalyticEvent(
-    route: HyperskillAnalyticRoute,
+class StepCompletionDailyStepCompletedModalShownHyperskillAnalyticEvent(
+    route: HyperskillAnalyticRoute
 ) : HyperskillAnalyticEvent(
     route,
     HyperskillAnalyticAction.SHOWN,
-    HyperskillAnalyticPart.NOTICE,
-    HyperskillAnalyticTarget.DAILY_NOTIFICATIONS_NOTICE
+    HyperskillAnalyticPart.MODAL,
+    HyperskillAnalyticTarget.DAILY_STEP_COMPLETED_MODAL
 )
